Fixed the encoding for REAL and worked around Update not updating files.
Since Sublime doesn't forcefully update files (after doing a git update
for example), we don't get the new data in the editor and it is thus very
possible to overwrite other people's commits (which we did some too many
times).

We're now instead creating a new scratch view. When switching back to another
file Sublime will automatically update it and save the world!